requirements.txt
# Core LLM and AI libraries
openai>=1.12.0
anthropic>=0.18.0
transformers>=4.38.0
torch>=2.2.0
accelerate>=0.27.0
bitsandbytes>=0.42.0

# PEFT and Fine-tuning
peft>=0.8.0
trl>=0.7.10
datasets>=2.17.0

# RAG and Vector Databases
langchain>=0.1.7
langchain-community>=0.0.20
chromadb>=0.4.22
faiss-cpu>=1.7.4
sentence-transformers>=2.3.1

# Document Processing
pypdf>=4.0.1
python-docx>=1.1.0
pytesseract>=0.3.10

# Web and APIs
requests>=2.31.0
beautifulsoup4>=4.12.3
fastapi>=0.109.2
uvicorn>=0.27.1

# Utilities
python-dotenv>=1.0.1
pydantic>=2.6.1
numpy>=1.26.4
pandas>=2.2.0
tqdm>=4.66.1

# Evaluation
rouge-score>=0.1.2
bert-score>=0.3.13
nltk>=3.8.1

# Development
jupyter>=1.0.0
pytest>=8.0.0
black>=24.1.1
flake8>=7.0.0

# Visualization
matplotlib>=3.8.2
seaborn>=0.13.2
